made new app for quiz separate from users
