+++ Quiz/urls.py
@@ -20,5 +20,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('users.urls'))
+    path('users/', include('users.urls')),
+    path('', include('quizapp.urls')),
+    
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
