+++ quizapp/templates/quizapp/results.html
@@ -2,6 +2,7 @@
 {% load index %}
 {% block content %}     
 
+<!-- Shows the Result of the Quiz attempted by the user after submitting -->
     <div class="container">
             <div class="row">
                 <div class="col border p-2">
