+++ quizapp/templates/quizapp/take_quiz.html
@@ -18,24 +18,14 @@
 
     <!-- Option 1: Bootstrap Bundle with Popper -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
     
     <script src="{% static 'main.js' %}" defer > </script>
 
-    {% if messages %}
-      {% for message in messages %}
-        <div class="alert alert-{{ message.tags }}">
-          {{message}}
-        </div>
-      {% endfor%}
-    {% endif %}
-
-
     <form method="POST" action="{% url 'results' %}">
         {% csrf_token %}
+
         <input type="text" hidden="true" id="quiz_id1" name="quiz_id" value="{{quiz_id}}"/>
         <input type="text" hidden="true" id="question_id" name="question_id" value="{{question.question_id}}"/>
         <input type="text" hidden="true" id="question_type" name="question_type" value="{{question.type}}"/>
@@ -104,31 +94,32 @@
     
 
 
-<script src="https://code.jquery.com/jquery-3.6.0.js" 
-    integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk=" 
-    crossorigin="anonymous"></script>
-    <script>
+    <script src="https://code.jquery.com/jquery-3.6.0.js" 
+        integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk=" 
+        crossorigin="anonymous"></script>
 
-        
-        
+    <script>
     
         $(document).ready(function() {
-
+            // By default show the submit button 
             $('#submit').show();
             $('#next').hide();
             var last = $('#last').val();
             var q_type = $('#question_type').val();
             if (q_type == 'MCQ'){
+                // show the MCQ Box when the question type is MCQ
                 $('#fib_box').hide();
                 $('#mcq_box').show();
             }
             else
             {
+                // show the FIB Box when the question type is FIB
                 $('#fib_box').show();
                 $('#mcq_box').hide();
             }
-
-            if (last == 'yes'){
+                // If it is last question then show 
+                // the submit button else hide it
+            if (last == 'yes'){   
                 $('#next').hide();
                 $('#submit').show();
             }
@@ -139,6 +130,7 @@
         });
 
         // when the user closes the tab or refreshes the page 
+        // save the quiz and submit it with blank answers
         $('#save_and_submit').click(function(e){
             var time_rem = $('#time_remaining').text();
             let q_id = $('#quiz_id1').val();
@@ -159,7 +151,7 @@
                 }
             }
             var mydata = {quiz_id:q_id, question_id:ques_id, choice:ch, time_remaining:time_rem, csrfmiddlewaretoken:csrf, last:last};
-            
+            // Do ajax post to results
             $.ajax({
             url: '{% url "results" %}',
             method: 'POST',
@@ -171,7 +163,7 @@
             });
         });
 
-        // when the user clicks on save and logout
+        // when the user clicks on "save and logout"
         $('#save').click(function(e){
             e.preventDefault();
             var time_rem = $('#time_remaining').text();
@@ -179,6 +171,8 @@
             let ques_id = $('#question_id').val();
             let csrf = $('input[name=csrfmiddlewaretoken]').val();
             var last = $('#last').val();
+            // If the question is last then submit the quiz
+            // else save it and logout
             if (last == 'no'){
                 if ($('#question_type').val()=='MCQ'){
                     var ch = $('input:radio[name=btnradio]:checked').val();
@@ -203,7 +197,6 @@
                         top.location.href="{% url 'logout' %}"
                     }
                 });
-                
             }
             else{
                 $('#submit').click();
@@ -213,7 +206,6 @@
         // when the user clicks on next button and gets the next question
         $('#next').click(function(e){
             e.preventDefault();
-            alert('next clicked')
             let q_id = $('#quiz_id1').val();
             let ques_id = $('#question_id').val();
             let csrf = $('input[name=csrfmiddlewaretoken]').val();
@@ -221,7 +213,7 @@
                 var ch = $('input:radio[name=btnradio]:checked').val();
                 if (ch==undefined){
                     ch = ''
-                }
+                    }
                 }
             else{
                 var ch = $('#fib_answer').val();
@@ -286,11 +278,7 @@
         });
 
   </script>
-   
-
 </body>
-
-    
 </html>
