--- users/templates/users/profile.html
@@ -1,29 +1,0 @@
-{% extends "base.html" %}
-{% load index %}
-{% block content %}
-
-    <div class="container">
-
-        {% for quiz in taken %}
-            <div class="row border m-2">
-                <div class="row m-2 p-1">
-                    <h4>{{quiz.quiz_id.quiz_name}}</h4>
-                </div>
-                <div class="row m-2 p-1">
-                    <h5>{{quiz.quiz_id.quiz_desc}}</h5>
-                </div>
-
-                <div class="row m-2 p-1">
-                    <form action="{% url 'profile' %}" method="POST">
-                        {% csrf_token %}
-                        <input type="hidden" name="quiz_id" value="{{quiz.quiz_id.quiz_id}}"/>
-                        <button id="show_details" class="btn btn-primary"> Show Details </button>
-                    </form>
-                </div> 
-
-            </div>
-        {% endfor %}
-    </div>
-
-{% endblock content %}
-
