+++ users/templates/users/register.html
@@ -1,6 +1,7 @@
 {% extends "base.html" %}
 {% load crispy_forms_tags %}
 {% block content %}
+
     <script type="text/javascript">
         function preback(){
             window.history.forward();
